+++ src/components/Card/cardChangelogData.ts
@@ -10,6 +10,13 @@ import { ChangelogData } from "../../utils/ComponentChangelogTable";
 
 export const changelogData: ChangelogData[] = [
   {
+    date: "2024-09-19",
+    version: "3.3.2",
+    type: "Update",
+    affects: ["Functionality"],
+    notes: ["Adds `fallbackSrc` and `onError` to the `imageProps` prop."],
+  },
+  {
     date: "2024-05-09",
     version: "3.1.2",
     type: "Update",