+++ src/components/FeedbackBox/feedbackBoxChangelogData.ts
@@ -10,6 +10,16 @@ import { ChangelogData } from "../../utils/ComponentChangelogTable";
 
 export const changelogData: ChangelogData[] = [
   {
+    date: "2024-09-19",
+    version: "3.3.2",
+    type: "Update",
+    affects: ["Styles"],
+    notes: [
+      "Changes 'r' in '(required)' label from upper- to lowercase",
+      "Increases the z-index of the `Drawer` component so it displays above all other elements on a page.",
+    ],
+  },
+  {
     date: "2024-07-25",
     version: "3.2.0",
     type: "Update",